[CHORE] API call 호출 로직 디렉토리 변경
