+++ src/apis/ootd/PostLikeService.ts
@@ -1,19 +1,13 @@
 import { authAxiosInstance } from '@/apis/utils'
-import { AxiosError, type AxiosResponse } from 'axios'
-import type { FollowerPageResponse, FollowerResponse } from '@/services/ootd/FollowDto'
+import type { PostPageResponse, PostResponse } from '@/apis/ootd/PostDto'
+import type { AxiosResponse } from 'axios'
+import { AxiosError } from 'axios'
 
 const SNS_SERVICE_PREFIX_PATH = '/sns-service'
 
-export const getFollowers = async (page: number, size: number, sort: string): Promise<FollowerPageResponse<FollowerResponse>> => {
+export const togglePostLike = async (postId: number): Promise<void> => {
   try {
-    const response: AxiosResponse = await authAxiosInstance.get(`/follows/followers`, {
-      params: {
-        page: page,
-        size: size,
-        sort: sort
-      }
-    })
-    return response.data
+    await authAxiosInstance.put(`/posts/${postId}/likes`)
   } catch (error) {
     if (error instanceof AxiosError) {
       if (error.response) {
